pyproject.toml
[tool.poetry]
name = "karros-backend"
version = "0.1.0"
description = ""
authors = ["catalunha <catalunha.mj@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12"
django = "^5.0.7"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^10.4.0"
python-decouple = "^3.8"
dj-database-url = "^2.2.0"
django-cors-headers = "^4.4.0"
psycopg2-binary = "^2.9.9"
gunicorn = "^22.0.0"
django-storages = {extras = ["s3"], version = "^1.14.4"}
drf-spectacular = "^0.27.2"
boto3 = "^1.34.153"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
